requirements.txt
pandas
kafka-python
pyspark
psycopg2-binary
pymongo
requests
pytest